enable use of typed values from javascript expressions
