+++ backend/molgenis-emx2-sql/src/main/resources/org/molgenis/emx2/sql/migration12.sql
