+++ backend/molgenis-emx2/src/main/java/org/molgenis/emx2/utils/JavaScriptUtils.java
@@ -19,11 +19,20 @@ public class JavaScriptUtils {
     // hide constructor
   }
 
-  public static String executeJavascript(String script) {
-    return executeJavascriptOnMap(script, null);
+  public static Object executeJavascript(String script) {
+    return executeJavascript(script, Object.class);
   }
 
-  public static String executeJavascriptOnMap(String script, Map<String, Object> values) {
+  public static Object executeJavascript(String script, Class clazz) {
+    return executeJavascriptOnMap(script, null, clazz);
+  }
+
+  public static Object executeJavascriptOnMap(String script, Map<String, Object> values) {
+    return executeJavascriptOnMap(script, values, Object.class);
+  }
+
+  public static Object executeJavascriptOnMap(
+      String script, Map<String, Object> values, Class clazz) {
     try {
       final Context context =
           Context.newBuilder("js")
@@ -42,7 +51,7 @@ public class JavaScriptUtils {
         }
       }
       String scriptWithFixedRegex = script.replace("\\\\", "\\");
-      return context.eval("js", scriptWithFixedRegex).toString();
+      return context.eval("js", scriptWithFixedRegex).as(clazz);
     } catch (Exception e) {
       throw new MolgenisException("script failed: " + e.getMessage());
     }
