+++ backend/molgenis-emx2/src/main/java/org/molgenis/emx2/utils/TypeUtils.java
@@ -15,8 +15,7 @@ import java.util.stream.Stream;
 import org.jooq.DataType;
 import org.jooq.JSONB;
 import org.jooq.impl.SQLDataType;
-import org.molgenis.emx2.ColumnType;
-import org.molgenis.emx2.MolgenisException;
+import org.molgenis.emx2.*;
 
 public class TypeUtils {
   private static final String LOOSE_PARSER_FORMAT =
@@ -477,6 +476,89 @@ public class TypeUtils {
       return LocalDateTime.ofInstant(Instant.ofEpochMilli(milliseconds), ZoneId.systemDefault());
     } else {
       return null;
+    }
+  }
+
+  public static Iterable<Row> convertToRows(TableMetadata metadata, List<Map<String, Object>> map) {
+    List<Row> rows = new ArrayList<>();
+    for (Map<String, Object> object : map) {
+      Row row = new Row();
+      for (Column column : metadata.getColumns()) {
+        if (object.containsKey(column.getIdentifier())) {
+          if (column.isRef()) {
+            convertRefToRow((Map<String, Object>) object.get(column.getIdentifier()), row, column);
+          } else if (column.isReference()) {
+            // REFBACK, REF_ARRAY
+            convertRefArrayToRow(
+                (List<Map<String, Object>>) object.get(column.getIdentifier()), row, column);
+          } else if (column.isFile()) {
+            BinaryFileWrapper bfw = (BinaryFileWrapper) object.get(column.getIdentifier());
+            if (bfw == null || !bfw.isSkip()) {
+              // also necessary in case of 'null' to ensure all file metadata fields are made empty
+              // skip is used when use submitted only metadata (that they received in query)
+              row.setBinary(
+                  column.getName(), (BinaryFileWrapper) object.get(column.getIdentifier()));
+            }
+          } else {
+            row.set(column.getName(), object.get(column.getIdentifier()));
+          }
+        }
+      }
+      rows.add(row);
+    }
+    return rows;
+  }
+
+  protected static void convertRefArrayToRow(
+      List<Map<String, Object>> list, Row row, Column column) {
+
+    List<Reference> refs = column.getReferences();
+    for (Reference ref : refs) {
+      if (!ref.isOverlapping()) {
+        if (!list.isEmpty()) {
+          row.set(ref.getName(), getRefValueFromList(ref.getPath(), list));
+        } else {
+          row.set(ref.getName(), new ArrayList<>());
+        }
+      }
+    }
+  }
+
+  private static List<Object> getRefValueFromList(
+      List<String> path, List<Map<String, Object>> list) {
+    List<Object> result = new ArrayList<>();
+    for (Map<String, Object> map : list) {
+      Object value = getRefValueFromMap(path, map);
+      if (value != null) {
+        result.add(value);
+      }
+    }
+    return result;
+  }
+
+  private static Object getRefValueFromMap(List<String> path, Map<String, Object> map) {
+    if (path.size() == 1) {
+      return map.get(path.get(0));
+    } else {
+      // should be > 1 and value should be of type map
+      Object value = map.get(path.get(0));
+      if (value != null) {
+        return getRefValueFromMap(path.subList(1, path.size()), (Map<String, Object>) value);
+      }
+      return null;
+    }
+  }
+
+  protected static void convertRefToRow(Map<String, Object> map, Row row, Column column) {
+    for (Reference ref : column.getReferences()) {
+      if (!ref.isOverlapping()) {
+        String name = ref.getName();
+        if (map == null) {
+          row.set(name, null);
+        } else {
+          row.set(ref.getName(), getRefValueFromMap(ref.getPath(), map));
+        }
+      }
     }
   }
 }
