+++ backend/molgenis-emx2/src/test/java/org/molgenis/emx2/utils/TestJavaScriptUtils.java
@@ -13,7 +13,7 @@ class TestJavaScriptUtils {
   @Test
   void testCalculateComputedExpression() {
     String expression = "5 + 7";
-    String outcome = executeJavascriptOnMap(expression, Map.of());
-    assertEquals(12, Integer.parseInt(outcome));
+    Object outcome = executeJavascriptOnMap(expression, Map.of());
+    assertEquals(12, outcome);
   }
 }
